+++ src/main/resources/sql/courseCreate.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS course
 (
-    course_id     integer                NOT NULL,
+    course_id     bigint                 NOT NULL,
     title         character varying(255) NOT NULL,
     creation_date timestamp              NOT NULL,
     last_modified timestamp              NOT NULL,
@@ -9,13 +9,13 @@ CREATE TABLE IF NOT EXISTS course
 );
 CREATE TABLE IF NOT EXISTS course_theme
 (
-    course_theme_id     integer                NOT NULL,
+    course_theme_id     bigint                 NOT NULL,
     title               character varying(255) NOT NULL,
     last_modified       timestamp              NOT NULL,
     description         character varying(255) NOT NULL,
     content             text                   NOT NULL,
     course_theme_number integer                NOT NULL,
-    course_id           integer                NOT NULL,
+    course_id           bigint                 NOT NULL,
     FOREIGN KEY (course_id) REFERENCES course (course_id) ON DELETE CASCADE,
     PRIMARY KEY (course_theme_id)
 );
